Adding options to menu and changing int rate to double

+++ src/NTNA1.java
@@ -5,24 +5,24 @@ public class NTNA1 extends Precision {
 	public NTNA1() {
 	}
 	
-	public double calculateNetValue(double emissionValue, int percentualAmortizacao) {
+	public double calculateNetValue(double emissionValue, double percentualAmortizacao) {
 		double netValue = (emissionValue * percentualAmortizacao)/100;
 		return netValue;
 	}
 	
-	public double calculatePU(double quotationDolarA, double quotationDolarB, double emissionValue, int amortizationPercentege) {
+	public double calculatePU(double quotationDolarA, double quotationDolarB, double emissionValue, double amortizationPercentege) {
 		double puAmortization = (quotationDolarA/quotationDolarB) * calculateNetValue(emissionValue, amortizationPercentege);
 		setPUAmortization(formactNumber(puAmortization, 6, false, false));
 		return formactNumber(puAmortization, 6, false, false);
 	}
 	
-	public double calculateInterestFactor(String dtp, String dtup, int rate) {
+	public double calculateInterestFactor(String dtp, String dtup, double rate) {
 		Date date = new Date();
 		double factor = (date.formactDate(dtp, dtup)/360.0) * (rate/100.0);
 		return formactNumber(factor, 8, true, false);
 	}
 	
-	public double calculateInterest(String dtp, String dtup, int rate) {
+	public double calculateInterest(String dtp, String dtup, double rate) {
 		double interest = getPUAmortization() * calculateInterestFactor(dtp, dtup, rate);
 		return formactNumber(interest, 6, false, false); 	
 	}
